requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
cx-Freeze==5.1.1
fabric==2.4.0
future==0.16.0
getdevinfo==1.0.3
gevent==1.3.6
greenlet==0.4.15
idna==2.7
ifaddr==0.1.4
invoke==1.2.0
ipaddress==1.0.22
lxml==4.2.5
macholib==1.11
netifaces==0.10.7
parallel-ssh==1.8.1
paramiko==2.4.2
pbr==4.2.0
pefile==2018.8.8
prettytable==0.7.2
pssh==2.3.1
psutil==5.4.7
py-cpuinfo==4.0.0
pyasn1==0.4.4
pycparser==2.18
PyInstaller==3.4
PyNaCl==1.2.1
python-hwinfo==0.1.6
requests==2.19.1
six==1.11.0
ssh2-python==0.15.0.post9
stevedore==1.29.0
uptime==3.0.1
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
